Working useSubscription (but without delta query)
useSubscription can be called with any graphql query. The query is modified
before actually doing the actual subscription:
1. operation is changed to subscription
2. a startDate condition is added to where
3. startDate parameter or name of startDate variabkle could be provided in a
config object together with configuration of the name of the timestampField to
check and the name of idField
4. query result is limited to 1 element
5. selection set is limited to __typename and id

This subscription query will receive 1 element after which the same query will
be run again but now as an actual query and collecting all elements created
since startDate. Then, the subscription will be renewed with the startDate of
the last element received.